+++ Project Gaia/Assets/Scripts/GameSettings.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6fe32994cd41972409ec3019314025ba
+guid: 2d054e3006ed93a428a1405d97b65bb8
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
