--- Project Gaia/Assets/Scripts/PauseMenu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2d054e3006ed93a428a1405d97b65bb8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
